refactor: move expressions out of root folder
