+++ src/expressions/condition.test.ts
@@ -1,9 +1,10 @@
-import { getFilterExpression, FilterInput } from './filter';
+import { ConditionInput } from 'dynoexpr';
+import { getConditionExpression } from './condition';
 
-describe(`filter expression`, () => {
-  it(`builds the FilterExpression and NameValueMaps - comparison operators`, () => {
+describe(`condition expression`, () => {
+  it(`builds the ConditionExpression and NameValueMaps - comparison operators`, () => {
     expect.assertions(1);
-    const Filter = {
+    const Condition = {
       a: `foo`,
       b: `> 1`,
       c: `>= 2`,
@@ -13,10 +14,10 @@ describe(`filter expression`, () => {
       g: `BETWEEN 6 AND 7`,
       h: `IN (foo, bar)`,
     };
-    const params: FilterInput = { Filter };
-    const result = getFilterExpression(params);
+    const params: ConditionInput = { Condition };
+    const result = getConditionExpression(params);
     const expected = {
-      FilterExpression: [
+      ConditionExpression: [
         `#n2661 = :va4d8`,
         `#n578f > :v849b`,
         `#n5f33 >= :v862c`,
@@ -53,9 +54,9 @@ describe(`filter expression`, () => {
     expect(result).toStrictEqual(expected);
   });
 
-  it(`builds the FilterExpression and NameValueMaps - function`, () => {
+  it(`builds the ConditionExpression and NameValueMaps - function`, () => {
     expect.assertions(1);
-    const Filter = {
+    const Condition = {
       a: `attribute_exists`,
       b: `attribute_not_exists`,
       c: `attribute_type(S)`,
@@ -63,10 +64,10 @@ describe(`filter expression`, () => {
       e: `contains(foo)`,
       f: `size > 10`,
     };
-    const params: FilterInput = { Filter };
-    const result = getFilterExpression(params);
+    const params: ConditionInput = { Condition };
+    const result = getConditionExpression(params);
     const expected = {
-      FilterExpression: [
+      ConditionExpression: [
         `attribute_exists(#n2661)`,
         `attribute_not_exists(#n578f)`,
         `attribute_type(#n5f33,:v546e)`,
@@ -93,23 +94,26 @@ describe(`filter expression`, () => {
     expect(result).toStrictEqual(expected);
   });
 
-  it(`builds the FilterExpression and NameValueMaps - mixed operators`, () => {
+  it(`builds the ConditionExpression and NameValueMaps - mixed operators`, () => {
     expect.assertions(1);
-    const Filter = {
+    const Condition = {
       a: 1,
       b: `between 2 and 3`,
       c: `size > 4`,
     };
-    const params: FilterInput = { Filter };
-    const result = getFilterExpression(params);
+    const params: ConditionInput = {
+      Condition,
+      ConditionLogicalOperator: 'OR',
+    };
+    const result = getConditionExpression(params);
     const expected = {
-      FilterExpression: [
+      ConditionExpression: [
         `#n2661 = :v849b`,
         `#n578f between :v862c and :vbaf3`,
         `size(#n5f33) > :v122c`,
       ]
         .map((exp) => `(${exp})`)
-        .join(` AND `),
+        .join(` OR `),
       ExpressionAttributeNames: {
         '#n2661': `a`,
         '#n578f': `b`,
@@ -120,6 +124,29 @@ describe(`filter expression`, () => {
         ':v862c': 2,
         ':vbaf3': 3,
         ':v122c': 4,
+      },
+    };
+    expect(result).toStrictEqual(expected);
+  });
+
+  it(`builds the ConditionExpression and NameValueMaps - avoid erroring values map`, () => {
+    expect.assertions(1);
+    const Condition = {
+      a: `attribute_exists`,
+      b: `attribute_not_exists`,
+    };
+    const params: ConditionInput = { Condition };
+    const result = getConditionExpression(params);
+    const expected = {
+      ConditionExpression: [
+        `attribute_exists(#n2661)`,
+        `attribute_not_exists(#n578f)`,
+      ]
+        .map((exp) => `(${exp})`)
+        .join(` AND `),
+      ExpressionAttributeNames: {
+        '#n2661': `a`,
+        '#n578f': `b`,
       },
     };
     expect(result).toStrictEqual(expected);
