+++ src/expressions/condition.ts
@@ -1,23 +1,9 @@
+import { ConditionInput, ConditionOutput } from 'dynoexpr';
 import {
   buildConditionAttributeNames,
   buildConditionAttributeValues,
   buildConditionExpression,
-  LogicalOperator,
-} from './condition-helpers';
-import { DynamoDbValue } from './helpers';
-
-export type Condition = Record<string, DynamoDbValue>;
-export type ConditionInput = Partial<{
-  Condition?: Condition;
-  ConditionLogicalOperator: LogicalOperator;
-  ExpressionAttributeNames: { [key: string]: string };
-  ExpressionAttributeValues: { [key: string]: DynamoDbValue };
-}>;
-export type ConditionOutput = Partial<{
-  ConditionExpression: string;
-  ExpressionAttributeNames: { [key: string]: string };
-  ExpressionAttributeValues: { [key: string]: DynamoDbValue };
-}>;
+} from './helpers';
 
 type GetConditionExpressionFn = (params?: ConditionInput) => ConditionOutput;
 export const getConditionExpression: GetConditionExpressionFn = (
