+++ src/expressions/filter.test.ts
@@ -1,9 +1,10 @@
-import { getConditionExpression, ConditionInput } from './condition';
+import { FilterInput } from 'dynoexpr';
+import { getFilterExpression } from './filter';
 
-describe(`condition expression`, () => {
-  it(`builds the ConditionExpression and NameValueMaps - comparison operators`, () => {
+describe(`filter expression`, () => {
+  it(`builds the FilterExpression and NameValueMaps - comparison operators`, () => {
     expect.assertions(1);
-    const Condition = {
+    const Filter = {
       a: `foo`,
       b: `> 1`,
       c: `>= 2`,
@@ -13,10 +14,10 @@ describe(`condition expression`, () => {
       g: `BETWEEN 6 AND 7`,
       h: `IN (foo, bar)`,
     };
-    const params: ConditionInput = { Condition };
-    const result = getConditionExpression(params);
+    const params: FilterInput = { Filter };
+    const result = getFilterExpression(params);
     const expected = {
-      ConditionExpression: [
+      FilterExpression: [
         `#n2661 = :va4d8`,
         `#n578f > :v849b`,
         `#n5f33 >= :v862c`,
@@ -53,9 +54,9 @@ describe(`condition expression`, () => {
     expect(result).toStrictEqual(expected);
   });
 
-  it(`builds the ConditionExpression and NameValueMaps - function`, () => {
+  it(`builds the FilterExpression and NameValueMaps - function`, () => {
     expect.assertions(1);
-    const Condition = {
+    const Filter = {
       a: `attribute_exists`,
       b: `attribute_not_exists`,
       c: `attribute_type(S)`,
@@ -63,10 +64,10 @@ describe(`condition expression`, () => {
       e: `contains(foo)`,
       f: `size > 10`,
     };
-    const params: ConditionInput = { Condition };
-    const result = getConditionExpression(params);
+    const params: FilterInput = { Filter };
+    const result = getFilterExpression(params);
     const expected = {
-      ConditionExpression: [
+      FilterExpression: [
         `attribute_exists(#n2661)`,
         `attribute_not_exists(#n578f)`,
         `attribute_type(#n5f33,:v546e)`,
@@ -93,26 +94,23 @@ describe(`condition expression`, () => {
     expect(result).toStrictEqual(expected);
   });
 
-  it(`builds the ConditionExpression and NameValueMaps - mixed operators`, () => {
+  it(`builds the FilterExpression and NameValueMaps - mixed operators`, () => {
     expect.assertions(1);
-    const Condition = {
+    const Filter = {
       a: 1,
       b: `between 2 and 3`,
       c: `size > 4`,
     };
-    const params: ConditionInput = {
-      Condition,
-      ConditionLogicalOperator: 'OR',
-    };
-    const result = getConditionExpression(params);
+    const params: FilterInput = { Filter };
+    const result = getFilterExpression(params);
     const expected = {
-      ConditionExpression: [
+      FilterExpression: [
         `#n2661 = :v849b`,
         `#n578f between :v862c and :vbaf3`,
         `size(#n5f33) > :v122c`,
       ]
         .map((exp) => `(${exp})`)
-        .join(` OR `),
+        .join(` AND `),
       ExpressionAttributeNames: {
         '#n2661': `a`,
         '#n578f': `b`,
@@ -123,29 +121,6 @@ describe(`condition expression`, () => {
         ':v862c': 2,
         ':vbaf3': 3,
         ':v122c': 4,
-      },
-    };
-    expect(result).toStrictEqual(expected);
-  });
-
-  it(`builds the ConditionExpression and NameValueMaps - avoid erroring values map`, () => {
-    expect.assertions(1);
-    const Condition = {
-      a: `attribute_exists`,
-      b: `attribute_not_exists`,
-    };
-    const params: ConditionInput = { Condition };
-    const result = getConditionExpression(params);
-    const expected = {
-      ConditionExpression: [
-        `attribute_exists(#n2661)`,
-        `attribute_not_exists(#n578f)`,
-      ]
-        .map((exp) => `(${exp})`)
-        .join(` AND `),
-      ExpressionAttributeNames: {
-        '#n2661': `a`,
-        '#n578f': `b`,
       },
     };
     expect(result).toStrictEqual(expected);
